major update on dashbaord and minor on landing
